[chante] (UMCS-33) - remove invoice guaranteed example.
