+++ examples/InvoiceSecured/index.php
@@ -44,15 +44,63 @@ require_once __DIR__ . '/../../../../autoload.php';
 
 <body style="margin: 70px 70px 0;">
 
-<form id="payment-form" action="<?php echo CONTROLLER_URL; ?>" class="unzerUI form" novalidate>
+<p><a href="https://docs.unzer.com/docs/testdata" target="_blank">Click here to open our test data in new tab.</a></p>
+
+<form id="payment-form">
+    <div id="example-invoice-secured"></div>
+    <div id="customer" class="field">
+        <!-- The customer form UI element will be inserted here -->
+    </div>
+    <div class="field" id="error-holder" style="color: #9f3a38"></div>
     <button class="unzerUI primary button fluid" id="submit-button" type="submit">Pay</button>
 </form>
 
 <script>
     // Create an Unzer instance with your public key
-    // This is not actually needed for this example but we want the sandbox banner to show on the page.
     let unzerInstance = new unzer('<?php echo UNZER_PAPI_PUBLIC_KEY; ?>');
-</script
+
+    // Create an Invoice Secured instance
+    let InvoiceSecured = unzerInstance.InvoiceSecured();
+
+    // Create a customer instance and render the customer form
+    let Customer = unzerInstance.Customer();
+    Customer.create({
+        containerId: 'customer'
+    });
+
+    // Handle payment form submission.
+    let form = document.getElementById('payment-form');
+    form.addEventListener('submit', function(event) {
+        event.preventDefault();
+        let InvoiceSecuredPromise = InvoiceSecured.createResource();
+        let customerPromise = Customer.createCustomer();
+        Promise.all([InvoiceSecuredPromise, customerPromise])
+            .then(function(values) {
+                let paymentType = values[0];
+                let customer = values[1];
+                let hiddenInputPaymentTypeId = document.createElement('input');
+                hiddenInputPaymentTypeId.setAttribute('type', 'hidden');
+                hiddenInputPaymentTypeId.setAttribute('name', 'paymentTypeId');
+                hiddenInputPaymentTypeId.setAttribute('value', paymentType.id);
+                form.appendChild(hiddenInputPaymentTypeId);
+
+                let hiddenInputCustomerId = document.createElement('input');
+                hiddenInputCustomerId.setAttribute('type', 'hidden');
+                hiddenInputCustomerId.setAttribute('name', 'customerId');
+                hiddenInputCustomerId.setAttribute('value', customer.id);
+                form.appendChild(hiddenInputCustomerId);
+
+                form.setAttribute('method', 'POST');
+                form.setAttribute('action', '<?php echo CONTROLLER_URL; ?>');
+
+                // Submitting the form
+                form.submit();
+            })
+            .catch(function(error) {
+                $('#error-holder').html(error.customerMessage || error.message || 'Error')
+            })
+    });
+</script>
 
 </body>
 </html>
